refactor account with clap derive api
