+++ Cargo.toml
@@ -59,11 +59,11 @@ version = "0.2"
 version = "0.4.24"
 
 [dependencies.clap]
-version = "4.0"
-features = ["derive"]
+version = "4.4"
+features = ["derive", "wrap_help"]
 
 [dependencies.clap_complete]
-version = "4.0"
+version = "4.4"
 
 [dependencies.clap_mangen]
 version = "0.2"
