+++ src/completion/command.rs
@@ -1,10 +1,32 @@
-use clap::{value_parser, Parser};
+use anyhow::Result;
+use clap::{value_parser, CommandFactory, Parser};
 use clap_complete::Shell;
+use log::info;
+use std::io;
+
+use crate::{cli::Cli, printer::Printer};
 
 /// Print completion script for the given shell to stdout
 #[derive(Debug, Parser)]
-pub struct Generate {
+pub struct Command {
     /// Shell that completion script should be generated for
     #[arg(value_parser = value_parser!(Shell))]
     pub shell: Shell,
+}
+
+impl Command {
+    pub async fn execute(self, printer: &mut impl Printer) -> Result<()> {
+        info!("executing completion generate command");
+
+        let mut cmd = Cli::command();
+        let name = cmd.get_name().to_string();
+        clap_complete::generate(self.shell, &mut cmd, name, &mut io::stdout());
+
+        printer.print(format!(
+            "Shell script successfully generated for shell {}!",
+            self.shell
+        ))?;
+
+        Ok(())
+    }
 }
