+++ src/man/mod.rs
@@ -1,2 +1,1 @@
 pub mod command;
-pub mod handler;
